execute/requirements.txt
requests
python-dotenv
-e git+https://github.com/melon-yellow/py-wapp#egg=py-wapp